first-draft: Completed first draft of design
This change contains the first draft of the system design.

The diagram has been drawn in draw.io. Both XML source and exported
PNG files have been included. The latter is inlined within the
README.md file for convenience.

The bulk of the supporting documentation is contained within the
README.md file, along with guidance on contributing to the repo.

A couple of GitHub checks have been added to spell-check README.md
and validatate any XML changes committed to the repo.